requirements.txt
pandas
numpy
pycoingecko
datetime
requests>=2.20
pytrends
beautifulsoup4
Cryptory
python-binance
twilio
yfinance
pytz
alpaca-trade-api
ndg-httpsclient
